✅ More tests

+++ CC_retry/libft/ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: llourens <llourens@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2025/05/02 12:32:02 by llourens      #+#    #+#                 */
-/*   Updated: 2025/06/20 11:12:33 by lilo          ########   odam.nl         */
+/*   Updated: 2025/07/04 11:52:31 by lilo          ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@
 
 size_t	ft_strlen(const char *string)
 {
-	int	length;
+	size_t	length;
 
 	length = 0;
 	while (string[length])
